+++ src/interfaces/api/running_router.py
@@ -8,7 +8,7 @@ from src.interfaces.schema.running_schema import (
     OnewayRunningRequest,
     OnewayRunningResponse,
 )
-from src.route_engine.running_route_service import get_circular_route, get_oneway_route
+from src.route_engine.engines.running_route_service import get_circular_route, get_oneway_route
 
 router = APIRouter(
     prefix="/api/running",
