+++ src/repository/network/graph_repository.py
@@ -6,7 +6,7 @@ from sqlalchemy import select
 from src.database.postgresql import get_postgresql_db
 from src.entity.network.walk_edge import WalkEdge
 from src.entity.network.walk_node import WalkNode
-from src.route_engine.path_utils import remove_dead_ends
+from src.route_engine.engines.path_utils import remove_dead_ends
 
 
 class GraphRepository:
