+++ src/route_engine/engines/oneway/landmark.py
@@ -1,18 +1,18 @@
 import networkx as nx
-from src.route_engine.path_utils import extract_coordinates
+
+from src.route_engine.engines.path_utils import extract_coordinates
 
 
-def landmark_circular_route(
+def landmark_oneway_route(
     G: nx.Graph,
     start_node: int,
+    end_node: int,
     landmark_node: int,
     weight: str = "custom_score"
 ) -> dict:
     try:
-        # 1. start → landmark
         path1 = nx.shortest_path(G, start_node, landmark_node, weight=weight)
 
-        # 2. 구간 1 엣지에 페널티 (귀환 시 다른 길 유도)
         original_weights = {}
         for i in range(len(path1) - 1):
             u, v = path1[i], path1[i + 1]
@@ -20,10 +20,8 @@ def landmark_circular_route(
                 original_weights[(u, v)] = G[u][v][weight]
                 G[u][v][weight] = original_weights[(u, v)] * 100
 
-        # 3. landmark → start
-        path2 = nx.shortest_path(G, landmark_node, start_node, weight=weight)
+        path2 = nx.shortest_path(G, landmark_node, end_node, weight=weight)
 
-        # 4. 가중치 원복
         for (u, v), old_w in original_weights.items():
             G[u][v][weight] = old_w
 
@@ -40,7 +38,7 @@ def landmark_circular_route(
         }
 
     except nx.NetworkXNoPath:
-        print(f"경로 없음: {start_node} → {landmark_node} → {start_node}")
+        print(f"경로 없음: {start_node} → {landmark_node} → {end_node}")
         return {"nodes": [], "coordinates": [], "total_distance_km": 0.0}
     except Exception as e:
         print(f"오류: {e}")
