+++ src/route_engine/engines/route_flat.py
@@ -32,14 +32,14 @@ import folium
 from pydantic import BaseModel, Field
 
 from src.repository.network.graph_repository import GraphRepository
-from src.route_engine.path_utils import (
+from src.route_engine.engines.path_utils import (
     find_nearest_node,
     extract_coordinates,
     prune_dead_ends,
 )
-from src.route_engine.path_circular_random import random_walk_route
-from src.route_engine.path_oneway_dijkstra import dijkstra_route
-from src.route_engine.path_oneway_random import oneway_random_route
+from src.route_engine.engines.circular.random import random_walk_route
+from src.route_engine.engines.oneway.dijkstra import dijkstra_route
+from src.route_engine.engines.oneway.random import oneway_random_route
 
 # ═════════════════════════════════════════════════════════
 # 상수
