+++ src/route_engine/engines/running_route_service.py
@@ -22,10 +22,10 @@ from src.interfaces.schema.running_schema import (
     CourseInfo,
     OnewayRunningResponse,
 )
-from src.route_engine.path_circular_random import random_walk_route
-from src.route_engine.path_oneway_dijkstra import dijkstra_route
-from src.route_engine.path_oneway_random import oneway_random_route
-from src.route_engine.path_utils import (
+from src.route_engine.engines.circular.random import random_walk_route
+from src.route_engine.engines.oneway.dijkstra import dijkstra_route
+from src.route_engine.engines.oneway.random import oneway_random_route
+from src.route_engine.engines.path_utils import (
     extract_coordinates,
     find_nearest_node,
     prune_dead_ends,
