--- src/route_engine/features/base_feature.py
@@ -1,8 +1,0 @@
-from typing import Any
-
-def base_feature_binder(graph: Any, context: dict[str, Any] | None = None) -> Any:
-    """
-    기본 Feature 바인더 인터페이스입니다.
-    TODO: 팀원 공통 인터페이스를 정의합니다.
-    """
-    pass
