--- src/route_engine/features/feature_registry.py
@@ -1,19 +1,0 @@
-from typing import Any
-
-def register_features(graph: Any, requested_features: list[str]) -> Any:
-    """
-    요청된 feature 목록에 따라 적절한 provider를 호출하여 graph edge에 feature를 바인딩합니다.
-
-    표준 feature 이름: safety_score, cctv_density, police_accessibility,
-    nature_score, park_accessibility, river_accessibility,
-    slope_score, landmark_score, live_poi_accessibility
-
-    금지:
-    - custom_score 계산, profile weight 해석, circular/oneway engine 호출 금지
-    - FastAPI, Streamlit, 기존 src/service/route import 금지
-
-    TODO:
-    - feature 이름과 provider 함수의 매핑 테이블을 정의합니다.
-    - 알 수 없는 feature 이름에 대한 처리 정책을 정합니다.
-    """
-    pass
