--- src/route_engine/features/slope_features.py
@@ -1,12 +1,0 @@
-from typing import Any
-
-def bind_slope_features(graph: Any, context: dict[str, Any] | None = None) -> Any:
-    """
-    graph edge에 경사도(slope_score) 및 어린이보호구역(school_zone_score) feature를 바인딩합니다.
-
-    금지:
-    - custom_score 계산, route algorithm 호출, Streamlit/FastAPI import 금지
-
-    TODO: 기존 slope 관련 로직을 분석하여 함수 내부를 구현합니다.
-    """
-    pass
