--- src/route_engine/scoring/scoring_engine.py
@@ -1,12 +1,0 @@
-from typing import Any
-
-def calculate_custom_score(graph: Any, profile: dict) -> Any:
-    """
-    graph 각 edge의 feature와 profile 가중치를 곱해 custom_score를 계산하여 부여합니다.
-
-    금지:
-    - 경로 탐색(다익스트라 등) 알고리즘 실행 금지
-
-    TODO: 점수 계산 공식을 구현합니다.
-    """
-    pass
